[SPARK-43031][SS][CONNECT] Enable unit test and doctest for streaming
### What changes were proposed in this pull request?

Enable unit tests and doctests for streaming queries. A lot are skipped and needs to be un-skipped as the development goes on.

Note that I also separated the `{foreach, foreachBatch}` tests from the original test suite. Because currently they are not implemented in connect and it seems unnecessary to manually add a skip for all of them in `StreamingParityTests`. Also it doesn't hurt to separate them I think as these tests are large enough already.

### Why are the changes needed?

More tests is always better than less.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

It's test itself.

Closes #40691 from WweiL/SPARK-43031-connect-unittests.

Authored-by: Wei Liu <wei.liu@databricks.com>
