+++ python/pyspark/sql/dataframe.py
@@ -535,7 +535,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> with tempfile.TemporaryDirectory() as d:
         ...     # Create a table with Rate source.
         ...     df.writeStream.toTable(
-        ...         "my_table", checkpointLocation=d) # doctest: +ELLIPSIS
+        ...         "my_table", checkpointLocation=d)
         <...streaming.query.StreamingQuery object at 0x...>
         """
         return DataStreamWriter(self)
